fix : 최종 오류 수정

+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/comment/service/CommentService.java
@@ -17,6 +17,10 @@ import xyz.tomorrowlearncamp.newsfeed.domain.newsfeed.service.NewsFeedService;
 import xyz.tomorrowlearncamp.newsfeed.domain.user.entity.Users;
 import xyz.tomorrowlearncamp.newsfeed.domain.user.service.UsersService;
 import xyz.tomorrowlearncamp.newsfeed.global.exception.LoginUserException;
+import xyz.tomorrowlearncamp.newsfeed.global.exception.NotFoundCommentException;
+import xyz.tomorrowlearncamp.newsfeed.global.exception.UnauthorizedWriterException;
+
+import java.util.Optional;
 
 
 @Service
@@ -103,8 +107,8 @@ public class CommentService {
     public void deleteComment(Long commentId, Long userId) {
         Comment comment = commentRepository.findByIdOrElseThrow(commentId);
         // 세션 userId와 삭제하려는 댓글의 userId 비교
-        if (!userId.equals(comment.getUser().getId())) {
-            throw new LoginUserException();
+        if (!userId.equals(comment.getUser().getId()) || !userId.equals(comment.getNewsFeed().getUser().getId())) {
+            throw new UnauthorizedWriterException();
         }
         commentRepository.delete(comment);
     }
